Add log library to separate spdlog dependencies
