+++ include/cpu/instruction.h
@@ -1,8 +1,8 @@
 ﻿#ifndef INSTRUCTION_H
 #define INSTRUCTION_H
 
+#include "utils/log.h"
 #include "utils/pack.h"
-#include "utils/utils.h"
 #include <cstdint>
 #include <source_location>
