+++ src/app/parallel_rdp_wrapper.cpp
@@ -2,7 +2,7 @@
 #include "app/spirv.h"
 #include "mmio/vi.h"
 #include "rdp_device.hpp"
-#include "utils/utils.h"
+#include "utils/log.h"
 #include "wsi.hpp"
 
 namespace N64 {
