+++ src/cpu/cop1.cpp
@@ -1,5 +1,5 @@
 #include "cpu/cop1.h"
-#include "utils/utils.h"
+#include "utils/log.h"
 
 namespace N64 {
 namespace Cpu {
