+++ src/mmio/mi.cpp
@@ -1,6 +1,6 @@
 ﻿#include "mmio/mi.h"
 #include "n64_system/interrupt.h"
-#include "utils/utils.h"
+#include "utils/log.h"
 #include <cstdint>
 
 namespace N64 {
