+++ src/mmio/pi.cpp
@@ -3,7 +3,7 @@
 #include "mmio/mi.h"
 #include "n64_system/interrupt.h"
 #include "n64_system/scheduler.h"
-#include "utils/utils.h"
+#include "utils/log.h"
 
 namespace N64 {
 namespace Mmio {
