+++ src/mmio/si.cpp
@@ -2,7 +2,7 @@
 #include "memory/memory.h"
 #include "mmio/mi.h"
 #include "n64_system/interrupt.h"
-#include "utils/utils.h"
+#include "utils/log.h"
 
 namespace N64 {
 namespace Mmio {
