+++ src/mmu/CMakeLists.txt
@@ -3,7 +3,7 @@ target_sources(mmu PRIVATE
     mmu.cpp
     tlb.cpp
 )
-target_link_libraries(mmu PRIVATE
+target_link_libraries(mmu PUBLIC
     common
-    utils
+    log
 )
