+++ src/mmu/mmu.cpp
@@ -1,6 +1,6 @@
 ﻿#include "mmu/mmu.h"
 #include "mmu/tlb.h"
-#include "utils/utils.h"
+#include "utils/log.h"
 
 namespace N64 {
 namespace Mmu {
