+++ src/mmu/tlb.cpp
@@ -1,5 +1,5 @@
 #include "mmu/tlb.h"
-#include "utils/utils.h"
+#include "utils/log.h"
 
 namespace N64 {
 namespace Mmu {
