+++ src/n64_system/config.cpp
@@ -1,5 +1,5 @@
 ﻿#include "n64_system/config.h"
-#include "utils/utils.h"
+#include "utils/log.h"
 #include <iostream>
 #include <string_view>
