+++ src/rcp/CMakeLists.txt
@@ -2,7 +2,9 @@ add_library(rcp STATIC)
 target_sources(rcp PRIVATE
     rsp.cpp
 )
-target_link_libraries(rcp PRIVATE
+target_link_libraries(rcp PUBLIC
     common
-    utils
+    log
+    mmio
+    n64_system
 )
