+++ src/utils/CMakeLists.txt
@@ -2,9 +2,19 @@ add_library(utils STATIC)
 target_sources(utils PRIVATE
     utils.cpp
 )
-target_link_libraries(utils
-    PUBLIC spdlog
-    PRIVATE
+target_link_libraries(utils PUBLIC
+    common
+)
+
+add_library(log STATIC)
+target_sources(log PRIVATE
+    log.cpp
+)
+target_link_libraries(log PUBLIC
     common
     cpu
+    spdlog
+)
+target_precompile_headers(log PUBLIC
+    <spdlog/spdlog.h>
 )
